wip: abstract libraries
